change settings suma
